# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jewelry API",
        "description": "API for Jewelry Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/financial-management/discounts": {
            "get": {
                "tags": [
                    "Discounts"
                ],
                "summary": "Get all discounts",
                "description": "Retrieve a paginated list of all discounts",
                "operationId": "f3450ed9a20fd2e33a97ce8ace265d39",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discounts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Discounts"
                ],
                "summary": "Create a new discount",
                "description": "Create a new discount with code, type, and value",
                "operationId": "2b6844dd44c1c92d5973f96c32d3576c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code",
                                    "type",
                                    "value",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Summer Sale 2024"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "SUMMER2024"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "percentage",
                                            "fixed"
                                        ],
                                        "example": "percentage"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-31"
                                    },
                                    "min_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "max_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1000
                                    },
                                    "usage_limit": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Summer discount for all customers"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Discount created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "discount": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/financial-management/discounts/{id}": {
            "get": {
                "tags": [
                    "Discounts"
                ],
                "summary": "Get a specific discount",
                "description": "Retrieve detailed information about a specific discount",
                "operationId": "0447d22b02d4fdf325d7f6b12381d6dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Discount ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discount retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "discount": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Discount not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Discounts"
                ],
                "summary": "Update a discount",
                "description": "Update an existing discount's information",
                "operationId": "0f2d6c13ff7b281d81e9318949f1468c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Discount ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code",
                                    "type",
                                    "value",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Summer Sale 2024 Updated"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "SUMMER2024"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "percentage",
                                            "fixed"
                                        ],
                                        "example": "percentage"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-31"
                                    },
                                    "min_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "max_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2000
                                    },
                                    "usage_limit": {
                                        "type": "integer",
                                        "example": 200
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated summer discount"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Discount updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "discount": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Discount not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Discounts"
                ],
                "summary": "Delete a discount",
                "description": "Move a discount to trash (soft delete)",
                "operationId": "de7f5361317f19692968d60b6b6429a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Discount ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discount moved to trash successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource has been moved to trash."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Discount not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/options/discounts": {
            "get": {
                "tags": [
                    "Discounts"
                ],
                "summary": "Get active discounts for dropdown",
                "description": "Retrieve a list of active discounts for dropdown/select options",
                "operationId": "6165b825edc36cb8f82b981e986c7e4a",
                "responses": {
                    "200": {
                        "description": "Active discounts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "discounts": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Summer Sale 2024"
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "SUMMER2024"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "percentage"
                                                    },
                                                    "value": {
                                                        "type": "number",
                                                        "example": 15
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}